requirements.txt
django-bootstrap4==21.2
fontawesome-free==5.15.4
requests===2.27.1
python-dateutil===2.8.2
django-markdownify==0.9.0